requirements-dev.txt
-r requirements.txt
pytest>=8.0
pytest-cov>=5.0
pytest-flask>=1.3
black>=24.0
ruff>=0.5
mypy>=1.10
faker>=25.0
